remove old am_list binary
